refactor(st-09071): modularize skill activation runtime
